First pass at running without a config file
The strategy here is to populate cn=config from the config file, but not to utilise it at runtime.

Here's the background:

You have to let Debian manage the initial slapd config, because it attempts to start the daemon after package installation, which fails if configuration is not present. This failure breaks the rest of the Puppet run.

If slapd.d exists when the slaptest command is run to convert slapd.conf to slapd.d, no actual conversion takes place.
Because you now have a wrong configuration and a started slapd, you now have a wrong database. Both the slapd.d configuration and the /var/lib/ldap data must be purged before any new records can be added.

There's ONE BIG HUGE CAVEAT DANGER DANGER WARNING here that I've not yet figured out: when slapd.conf changes, it triggers a reset of slapd.d, which in turn wipes out the LDAP database. We can't go to production like this, obviously, but we can get a dev environment up and running.