Use ldapdn type for specifying individual config options
cnconfig_attrs was changed from a Hash to an Array, since keys may have multiple values in LDAP and this is the least-friction method of specification for those familiar with writing LDIFs.

This also obviated the dependency on stdlib, since we no longer need to check if cnconfig_attrs is empty.